add posts table for caching posts; handle amend messages; refactor node.py as a package
